pyproject.toml
[tool.poetry]
name = "ugahacksx"
version = "0.1.0"
description = ""
authors = ["apar2021 <apar2021@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pymongo = "^4.11"
flask = "^3.1.0"
flask-cors = "^5.0.0"
flask-pymongo = "^3.0.1"
scikit-learn = "^1.6.1"
jinja2 = "^3.1.5"
werkzeug = "^3.1.3"
openai = "^1.61.1"
python-dotenv = "^1.0.1"
flask-bcrypt = "^1.0.1"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
